tinybridge-backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.31.1
sqlalchemy==2.0.36
python-multipart==0.0.17
pydantic==2.10.3
python-dotenv==1.0.1